+++ src/list/list.common.ts
@@ -1,15 +1,14 @@
-import { Directive, TemplateRef } from '@angular/core'
+import { Directive, TemplateRef } from "@angular/core";
 
 export interface IListChild {
-	index: number;
+    index: number;
 }
 
 export enum IgxListPanState { NONE, LEFT, RIGHT }
 
-
 @Directive({
-	selector: "[igxEmptyList]"
+    selector: "[igxEmptyList]"
 })
 export class IgxEmptyListTemplateDirective {
-	constructor(public template: TemplateRef<any>) { }
+    constructor(public template: TemplateRef<any>) { }
 }