+++ src/list/list.component.ts
@@ -16,18 +16,15 @@ import {
 } from "@angular/core";
 
 import { IgxRippleModule } from "../directives/ripple/ripple.directive";
-import { IgxEmptyListTemplateDirective } from "./list.common"
-import { IgxListItemComponent } from "./list-item.component"
+
+import { IgxListItemComponent } from "./list-item.component";
+import { IgxEmptyListTemplateDirective } from "./list.common";
 
 // ====================== LIST ================================
 // The `<igx-list>` directive is a list container for items and headers
 @Component({
-	host: {
-		role: "list"
-	},
-	// changeDetection: ChangeDetectionStrategy.OnPush,
-	selector: "igx-list",
-	templateUrl: "list.component.html"
+    selector: "igx-list",
+    templateUrl: "list.component.html"
 })
 export class IgxListComponent {
 
@@ -105,9 +102,9 @@ export class IgxListComponent {
 }
 
 @NgModule({
-	declarations: [IgxListComponent, IgxListItemComponent, IgxEmptyListTemplateDirective],
-	exports: [IgxListComponent, IgxListItemComponent, IgxEmptyListTemplateDirective],
-	imports: [CommonModule, IgxRippleModule]
+    declarations: [IgxListComponent, IgxListItemComponent, IgxEmptyListTemplateDirective],
+    exports: [IgxListComponent, IgxListItemComponent, IgxEmptyListTemplateDirective],
+    imports: [CommonModule, IgxRippleModule]
 })
 export class IgxListModule {
 }